Made modifications to Reps.swift to make the model easier to use. Although I am finding some quirks with the Google Civic Information API. Namely, for some inexplicable reason, some Representatives or Senators do not have a photoURL field. Despite the fact that the API source for the image data (bioguide.congress.gov) does have valid images with valid URL's for them. I submitted an error to Google, we'll see what happens. For now, if the field is nil, it is filled in with "photoURL". Additionally, some lawmakers have a "line2" field for their office address. I will have to make considerations for this.
I created an IndividualLawmaker view which will display the info for individual lawmakers when selected. There's enough basic info displayed for now and I can start making mockups. MyRepsView is seemingly done for now.

For the MyRepsViewModel I added the checks for the photoURL field and made some slight changes to reduce redundancy in code.